Divided gui.py into separated modules
